Added workaround for svn tags imported wrongly
When some projects have migrated from Subversion to Git, there
were several tags that were treated as new commits, which shows
a change in the whole project (code added/removed) when nothing
really happened.  For instance, in GNOME a lot svn tags were
catched during the migration, but not all of them.

svn tags in git repositories brings bad stats because a commit
may involves several thousands of source of lines of code.

Signed-off-by: Germán Póo-Caamaño <gpoo@gnome.org>